+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0+5
+
+* onRender handlers added
+* onError, onPageError handlers added (Android only)
+* Add example snippet in readme
+
+## 1.0.0+4
+
+* Add controller functions description in document
 ## 1.0.0+3
 
 * Support `enableSwipe` in iOS